refactor: extract portfolio business logic to service layer (#153)
Move composite index computation and performer ranking from
routers/portfolio.py into services/portfolio.py. Router handlers
become thin wrappers (~10 lines each).
